application, blueprints 구조로 변경
프로젝트 구조를 app의 __init__.py에서
application, blueprints 파일을 사용하여 바꿈
