+++ app/account/__init__.py
@@ -1,5 +1,1 @@
-from flask import Blueprint
-
-account = Blueprint('account', __name__)
-
-from . import views, models, admin
+from . import models, admin
