+++ app/account/models.py
@@ -4,7 +4,7 @@ import os
 from flask_login import UserMixin
 from werkzeug.security import check_password_hash, generate_password_hash
 
-from app import db, login_manager
+from app.application import db, login_manager
 
 
 def generate_token():
