+++ app/account/views.py
@@ -1,8 +1,8 @@
 from flask import g, jsonify, request
 
-from app import db
+from app.application import db
 
-from . import account
+from app.blueprints import account
 from .forms import LoginForm, RegisterForm
 from .models import User
