+++ tests/conftest.py
@@ -1,7 +1,7 @@
 import pytest
 
-from app import db as _db
-from app import create_app
+from app.application import (create_app,
+                             db as _db)
 from app.account.models import User
